Changed int to long

+++ src/Main.java
@@ -1,6 +1,6 @@
 class Main {
     public static void main(String[] args) {
-        int[] saleManagerArray = {1, 5, 19};
+        long[] saleManagerArray = {1, 5, 19};
         SalesManager salesManager = new SalesManager(saleManagerArray);
 
         System.out.println(salesManager.max());
